Added support for screenshotting. Separated window code into window_utils
Author:   Evan Maicus <emaicus@outlook.com>
