+++ grading/main_runner.cpp
@@ -65,6 +65,7 @@ int main(int argc, char *argv[]) {
     if (my_testcase.isFileCheck()) continue;
     if (my_testcase.isCompilation()) continue;
     std::vector<std::string> commands = stringOrArrayOfStrings((*tc)[i],"command");
+    std::vector<std::string> actions  = stringOrArrayOfStrings((*tc)[i],"actions");
     assert (commands.size() > 0);
 
     std::cout << "TITLE " << my_testcase.getTitle() << std::endl;
@@ -87,6 +88,7 @@ int main(int argc, char *argv[]) {
       int exit_no = execute(commands[x] +
                             " 1>" + my_testcase.getPrefix() + "_STDOUT" + which + ".txt" +
                             " 2>" + my_testcase.getPrefix() + "_STDERR" + which + ".txt",
+                            actions,
                             logfile,
                             my_testcase.get_test_case_limits(),
                             config_json.value("resource_limits",nlohmann::json()),
@@ -106,7 +108,9 @@ int main(int argc, char *argv[]) {
         std::string filename     = my_testcase.getMyPrefixFilename(v,i);
         assert (raw_filename != "");
         if (access( raw_filename.c_str(), F_OK|R_OK|W_OK ) != -1) { // file exists
+          std::vector<std::string> actions;
           execute("/bin/mv "+raw_filename+" "+filename,
+                  actions,
                   "/dev/null",
                   my_testcase.get_test_case_limits(),
                   config_json.value("resource_limits",nlohmann::json()),
